double version conversion

+++ SimpleCalculator/ViewController.swift
@@ -15,27 +15,22 @@ class ViewController: UIViewController {
     @IBOutlet weak var secondText: UITextField!
     @IBOutlet weak var resultLabel: UILabel!
     
-    var result = 0
+    var result = 0.0
     
     
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view.
         
-    
-        
     }
     
     
     @IBAction func sumClicked(_ sender: Any) {
         
-        
-        if let firstNumber = Int(firstText.text!) {
-            if let secondNumber = Int(secondText.text!) {
-             
+        if let firstNumber = Double(firstText.text!) {
+            if let secondNumber = Double(secondText.text!) {
                 result = firstNumber + secondNumber
                 resultLabel.text = String(result)
-                
             }
         }
         
@@ -44,44 +39,38 @@ class ViewController: UIViewController {
     
     @IBAction func minusClicked(_ sender: Any) {
         
-        if let firstNumber = Int(firstText.text!) {
-            if let secondNumber = Int(secondText.text!) {
-                
+        if let firstNumber = Double(firstText.text!) {
+            if let secondNumber = Double(secondText.text!) {
                 result = firstNumber - secondNumber
                 resultLabel.text = String(result)
-                
             }
         }
+        
     }
     
     
     @IBAction func multiplyClicked(_ sender: Any) {
         
-        if let firstNumber = Int(firstText.text!) {
-            if let secondNumber = Int(secondText.text!) {
-                
+        if let firstNumber = Double(firstText.text!) {
+            if let secondNumber = Double(secondText.text!) {
                 result = firstNumber * secondNumber
                 resultLabel.text = String(result)
-                
             }
         }
+        
     }
     
     
     @IBAction func divideClicked(_ sender: Any) {
-        
-        if let firstNumber = Int(firstText.text!) {
-            if let secondNumber = Int(secondText.text!) {
-                
+    
+        if let firstNumber = Double(firstText.text!) {
+            if let secondNumber = Double(secondText.text!) {
                 result = firstNumber / secondNumber
                 resultLabel.text = String(result)
-                
             }
         }
         
     }
-    
-
 
 }
